Tweaked the UITests project to make it a little more helpful.
Made the SWFTestApp configuration a bit more verbose if it's configured wrong.  It should now throw helpful exceptions if the JavaLocation property isn't set to a valid path in App.config and if it can't find java.exe within the specified folder.

Additionally, added ProcessorType key into app.config as it requires this to identify which swt library it needs to add to the path for the java app when it attempts to start it.