Adding Maximum & Minimum properties to Slider
This commit introduces Maximum & Minimum properties to the Slider
control. Before this commit, there was no way to get the maximum value
for a Slider control. Because of this, it is not easy to slide to the
end from a test because we don't know when to stop calling
"LargeIncrementButton.Click()".