Better .DoExpand and Collapse for Win32TreeViews.
The original implementation was performing a DoubleClick on the item,
which often triggers some application-specific event.
Pressing the RIGHT and LEFT keys to expand-collapse trees is
a widely used idiom in applications and less likely to be
used for something else.

The exception message upon unsuccessful expanding hints that the
desired implementation would be to click on the expand button of
the tree item, yet I don't currently see a way to do that.
If it's not possible to do so, the error message should be
changed accordingly.