Extend Allotted Time to Expand ListControl Element
I hit an issue in my project where the 50ms to expand the combo box was not long enough. When reading Items.SelectedItemText, I was getting the incorrect value as the combo box was not fully expanded within the allotted time. This solution allows more time for slower combo boxes without adding any extra sleeping to combo boxes that don't need it.